Refine the code and decrease the parse method length in Parser
