+++ src/main/java/reboot/command/ErrorCommand.java
@@ -1,4 +1,5 @@
 package reboot.command;
+
 import reboot.Storage;
 import reboot.TaskList;
 import reboot.gui.Gui;
