+++ src/test/java/reboot/parser/ParserTest.java
@@ -1,4 +1,4 @@
-package reboot;
+package reboot.parser;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertTrue;
